feat(codegen): add MISRA 10.3 narrowing casts for bit extraction (#845)
When extracting bits to a narrower type (e.g., u32[0,8] -> u8), the generated
C code now includes explicit casts for MISRA C:2012 Rule 10.3 compliance.

- Integrates NarrowingCastHelper with PostfixExpressionGenerator
- Uses CodeGenState.expectedType to determine target type
- Applies getPromotedType() to handle C integer promotion
- Updates expected test outputs with new cast expressions

Before: uint8_t x = ((value >> 0U) & 0xFFU);
After:  uint8_t x = (uint8_t)((value >> 0U) & 0xFFU);

+++ tests/casting/bit-index-narrowing.expected.c
@@ -9,7 +9,7 @@
 // Using [start, width] syntax to explicitly extract bits
 int main(void) {
     uint32_t large = 0xDEADBEEFU;
-    uint8_t low_byte = ((large >> 0U) & 0xFFU);
-    uint8_t high_byte = ((large >> 24U) & 0xFFU);
-    uint16_t low_word = ((large >> 0U) & 0xFFFFU);
+    uint8_t low_byte = (uint8_t)((large >> 0U) & 0xFFU);
+    uint8_t high_byte = (uint8_t)((large >> 24U) & 0xFFU);
+    uint16_t low_word = (uint16_t)((large >> 0U) & 0xFFFFU);
 }
